+++ business/entries/src/main/kotlin/proton/android/authenticator/business/entries/domain/Entry.kt
@@ -23,6 +23,7 @@ import me.proton.core.crypto.common.keystore.EncryptedByteArray
 data class Entry(
     val id: String,
     val content: EncryptedByteArray,
+    val createdAt: Long,
     val modifiedAt: Long,
     val isDeleted: Boolean,
     val isSynced: Boolean,